some updates to notebook and code
