+++ recovar/em/core.py
@@ -1,0 +1,144 @@
+import functools
+import jax
+import jax.numpy as jnp
+from recovar import core
+import numpy as np
+import logging
+from scipy.spatial.transform import Rotation as R
+
+logger = logging.getLogger(__name__)
+
+
+# Probabilities will be 4 dimensional:
+IMAGE_AXIS=0
+VOL_AXIS=1
+ROT_AXIS=2
+TRANS_AXIS=3
+
+
+NORM_FFT = "backward"
+
+# batch volumes
+batch_vol_rot_slice_volume_by_map = jax.vmap(core.slice_volume_by_map, in_axes = (0, VOL_AXIS, None, None, None), out_axes=1 )
+batch_vol_slice_volume_by_map = jax.vmap(core.slice_volume_by_map, in_axes = (0, None, None, None, None), out_axes=1 )
+
+
+
+from recovar.fourier_transform_utils import fourier_transform_utils
+ftu = fourier_transform_utils(jnp)
+def crosscorr_from_ft(many_images, one_image, image_shape):
+    return ftu.get_idft2( jnp.conj(one_image.reshape(1, *image_shape)) * (many_images.reshape(-1, *image_shape)))#.reshape(-1, *image_shape)).reshape(-1, *image_shape)
+
+def norm_squared_residuals_from_ft_one_image(many_images, one_image, image_shape):
+    many_images_of_shape = many_images.shape
+    many_images = many_images.reshape(-1, many_images.shape[-1])
+    many_images = crosscorr_from_ft(many_images, one_image, image_shape)
+    many_images = many_images.reshape(many_images_of_shape)
+    return many_images
+
+norm_squared_residuals_from_ft = jax.vmap(norm_squared_residuals_from_ft_one_image, in_axes = (0, 0, None))
+
+
+
+@functools.partial(jax.jit, static_argnums=[4,6,8])
+def compute_dot_products(projections, batch, translations, CTF_params, CTF_fun, noise_variance, process_images, voxel_size, image_shape):
+    '''
+    Computes -2 * y_i.T @ (S_s C_i * Proj_j) for i,j,s
+    where C_i is CTF, S_s are shifts, and Proj_j are projections, y_i are batch (unprocessed)
+    '''
+    # proj_norm = jnp.linalg.norm(projections, axis=-1, keepdims=True)**2
+    # Technically don't need do to this step. Maybe should delete at some point.
+
+    batch = process_images(batch, apply_image_mask = False) 
+    batch_norm = jnp.linalg.norm(batch / jnp.sqrt(noise_variance), axis = (-1), keepdims = True)**2
+
+    batch *= CTF_fun( CTF_params, image_shape, voxel_size) / noise_variance
+    result = jnp.empty((batch.shape[0], projections.shape[0], translations.shape[0]), dtype = jnp.float32)
+
+    # One bad thing about the FFT shifting way is that is involve allocating size(projections) * num_images
+    # This way relies a lot on never allocating things of that size, so that we can handle a ton of projections x images at once.
+    # It's not very clear whether it's better to do. Compute all shifted images, then compute IP, or compute IP for each shift.
+    # The latter is probably better for memory, but the former is probably better for speed. We probably want to do the latter.
+
+    ## IMPLEMENATION 1
+    ## I thought this would avoid allocating all the shifted images at once, but it doesn't seem to be the case.
+    imp = 3
+    if imp == 1:
+        for i in range(translations.shape[0]):
+            # Probably should put a better fn here for shfiting but oh well.
+            shifted_batch = core.translate_images(batch, jnp.repeat(translations[i:i+1], batch.shape[0], axis=0), image_shape)
+
+            result = result.at[:,:,i].set(-2 * (jnp.conj(shifted_batch) @ projections.T).real + batch_norm)
+    
+    ## IMPLEMENATION 2
+    ## This doesn't work. I'm not sure if it should be any different than the typical for loop
+
+    # def body_fun(i, result):
+    #     shifted_batch = core.translate_images(batch, jnp.repeat(translations[i:i+1], batch.shape[0], axis=0), image_shape)
+
+    #     result = result.at[:,:,i].set(-2 * (jnp.conj(shifted_batch) @ projections.T).real + batch_norm)
+
+    # result = jax.lax.fori_loop(0, translations.shape[0], body_fun, result)
+
+    # IMPLEMENTATION 3. Just one big matvec.
+    # This is probably allocating far more memory than needed.
+    elif imp == 3:
+        shifted_images = core.batch_trans_translate_images(batch, jnp.repeat(translations[None], batch.shape[0], axis=0), image_shape)
+        n_shifted_images = np.prod(shifted_images.shape[:-1])
+        result = -2 * (jnp.conj(shifted_images).reshape(n_shifted_images, shifted_images.shape[-1] ) @ projections.T).real 
+        result = result.reshape(batch.shape[0], translations.shape[0], projections.shape[0]) + batch_norm[:,None]
+        result = result.swapaxes(1,2)
+        # May want to swap axes
+    
+
+    return result
+
+
+@functools.partial(jax.jit, static_argnums=[2,5])
+def compute_CTFed_proj_norms(projections, CTF_params, CTF_fun, noise_variance, voxel_size, image_shape):
+    '''
+    Computes  |C_i Proj_j|^2 for i,j by writing it as a mat-mat
+    where C_i is CTF, S_s are shifts, and Proj_j are projections
+    '''
+    CTFs = CTF_fun( CTF_params, image_shape, voxel_size)**2 / noise_variance
+
+    result = CTFs @ projections.T
+
+    return result
+
+
+
+
+def probabilities_to_hard_assignment_pose(probabilities, rotation_grid, translation_grid):
+    idx = np.argmax(probabilities.reshape(probabilities.shape[0], -1), axis=-1)
+    return hard_assignment_idx_to_pose(idx, rotation_grid, translation_grid)
+
+def probabilities_to_hard_assignment_idx(probabilities, rotation_grid, translation_grid):
+    idx = np.argmax(probabilities.reshape(probabilities.shape[0], -1), axis=-1)
+    square_shape = (rotation_grid.shape[0], translation_grid.shape[0])
+    maxpos_vect = np.column_stack(np.unravel_index(idx,square_shape))
+    rot_idx = maxpos_vect[:,0]
+    trans_idx = maxpos_vect[:,1]
+    return rot_idx, trans_idx
+
+def hard_assignment_idx_to_pose(indices, rotation_grid, translation_grid):
+    square_shape = (rotation_grid.shape[0], translation_grid.shape[0])
+    maxpos_vect = np.column_stack(np.unravel_index(indices,square_shape))
+    predicted_trans = translation_grid[maxpos_vect[:,1]]
+    predicted_pose = rotation_grid[maxpos_vect[:,0]]
+    return predicted_pose, predicted_trans
+
+
+def estimate_error_from_hard_assignment(hard_assignment, gt_pose, gt_trans, rotation_grid, translation_grid):
+    predicted_pose, predicted_trans = hard_assignment_idx_to_pose(hard_assignment, rotation_grid, translation_grid)
+    # gt_pose = rotation_grid[true_rot_indices]
+    predicted_pose = R.from_matrix(predicted_pose)
+    gt_pose = R.from_matrix(gt_pose)
+    error = (predicted_pose * gt_pose.inv()).magnitude() / np.pi * 180
+    
+    mean_angle_error = np.mean(error)
+    mean_trans_error = np.mean(np.linalg.norm(predicted_trans - gt_trans, axis=-1))
+    logger.info(f"mean trans error: {mean_trans_error} pixels")
+    logger.info(f"mean angle error: {mean_angle_error} degrees")
+
+    return np.mean(error), np.mean(np.linalg.norm(predicted_trans - gt_trans, axis=-1))
