+++ recovar/noise.py
@@ -686,9 +686,9 @@ def estimate_noise_level_no_masks(experiment_dataset, image_subset, mean_estimat
     rhs = 0 
     
     # Print debug info about input parameters
-    print(f"Image subset type: {type(image_subset)}")
-    if image_subset is not None:
-        print(f"Image subset shape: {image_subset.shape if hasattr(image_subset, 'shape') else 'no shape'}")
+    # print(f"Image subset type: {type(image_subset)}")
+    # if image_subset is not None:
+    #     print(f"Image subset shape: {image_subset.shape if hasattr(image_subset, 'shape') else 'no shape'}")
     
     data_generator = experiment_dataset.get_image_subset_generator(
         batch_size=batch_size,
